requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-redis==5.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
passlib==1.7.4
psycopg2==2.9.1
pytz==2021.1
redis==3.5.3
Shapely==1.7.1
sqlparse==0.4.1
whitenoise==5.2.0
